tests: move FakeApt fixtures into deb tests
These are no longer shared for other tests so we can freely
modify without breaking external tests.  We'll move them
into test_deb to discourage further external use.
